requirements.txt
asgiref==3.6.0
chardet==5.1.0
Django==4.1.7
django-crispy-forms==1.14.0
Pillow==9.0.1
pytils==0.4.1
sqlparse==0.4.2
tzdata==2021.5
gunicorn==20.1.0
django-heroku==0.3.1
